replace classify_message usage with the Deserializer
