+++ tests/test_subscribe.py
@@ -15,62 +15,6 @@ import pytest
 from hop import subscribe
 from hop.models import GCNCircular, VOEvent, MessageBlob
 
-
-def content_mock(message_model):
-    """Mock content to pass during the message_model creation since MagicMock()
-    is unable to mock __init__ of the model dataclass in tests.
-    """
-    content = {}
-    for field in fields(message_model):
-        content.update({field.name: "test"})
-    return content
-
-
-@pytest.mark.parametrize(
-    "message",
-    [
-        {"format": "voevent", "content": content_mock(VOEvent)},
-        {"format": "circular", "content": content_mock(GCNCircular)},
-        {"format": "blob", "content": content_mock(MessageBlob)},
-        {"format": "other", "content": "other"},
-        ["wrong_datatype"],
-        {"wrong_key": "value"},
-    ],
-)
-def test_classify_message(message, message_parameters_dict):
-
-    # test a non-dict message
-    if not isinstance(message, dict):
-        with pytest.raises(ValueError):
-            test_model = subscribe.classify_message(message)
-        return
-    # test a dict message with wrong key values
-    elif not (("format" in message) and ("content" in message)):
-        with pytest.raises(KeyError):
-            test_model = subscribe.classify_message(message)
-        return
-
-    message_format = message["format"]
-
-    # test an invalid format
-    if message_format not in message_parameters_dict:
-        with pytest.raises(ValueError):
-            test_model = subscribe.classify_message(message)
-        return
-
-    # load parameters from conftest for valid formats
-    message_parameters = message_parameters_dict[message_format]
-    model_name = message_parameters["model_name"]
-    expected_model = message_parameters["expected_model"]
-
-    # test valid formats
-    with patch(f"hop.models.{model_name}", MagicMock()):
-        test_model = subscribe.classify_message(message)
-
-    # verify the message is classified properly
-    assert isinstance(test_model, expected_model)
-
-
 # test the subscribe printer for each message format
 @pytest.mark.parametrize("message_format", ["voevent", "circular", "blob"])
 def test_print_message(message_format, message_parameters_dict):
